Small refactor with spawning working better
