+++ source/Missions/IMissionGameManager.cs
@@ -2,10 +2,10 @@
 using TaleWorlds.Library;
 using TaleWorlds.MountAndBlade;
 
-namespace Missions.Services.Network
+namespace Missions
 {
     public interface IMissionGameManager
     {
-        Agent SpawnAgent(Vec3 startingPos, CharacterObject character);
+        void StartGame();
     }
 }
